Fix bug with DensityGate when gate_fraction=0.
When `gate_fraction` is 0, `accepted_data_indices` is an empty numpy
array, which is an object array by default and later causes an error
when it's used to index another array. Fix this by ensuring
`accepted_data_indices` is always an integer array.

+++ FlowCal/gate.py
@@ -583,7 +583,8 @@ def density2d(data,
                                                dtype=np.int)]
     accepted_data_indices = np.array([item       # flatten list of lists
                                       for sublist in accepted_data_indices
-                                      for item in sublist])
+                                      for item in sublist],
+                                     dtype=np.int)
     accepted_data_indices = np.sort(accepted_data_indices)
 
     # Convert list of accepted data indices to boolean mask array
